web/requirements.txt
# Requirements for DriftApp Web
Django>=4.2
djangorestframework>=3.14
